+++ src/components/container.js
@@ -23,7 +23,7 @@ import {connect} from 'react-redux';
 import memoize from 'lodash.memoize';
 import {console as Console} from 'global/window';
 import {injector} from './injector';
-import KeplerGlFactory, {keplerGlChildDeps} from './kepler-gl';
+import KeplerGlFactory from './kepler-gl';
 import {forwardTo} from 'actions/action-wrapper';
 
 import {
@@ -116,13 +116,19 @@ export function ContainerFactory(KeplerGl) {
   return connect(mapStateToProps, dispatchToProps)(Container);
 }
 
-// provide all recipes to injector
-export const appInjector = [
-  ContainerFactory,
-  ...ContainerFactory.deps,
-  ...KeplerGlFactory.deps,
-  ...keplerGlChildDeps
-].reduce((inj, factory) => inj.provide(factory, factory), injector());
+// entryPoint
+function flattenDeps(allDeps, factory) {
+  const addToDeps = allDeps.concat([factory]);
+  return Array.isArray(factory.deps) && factory.deps.length ?
+    factory.deps.reduce((accu, dep) => flattenDeps(accu, dep), addToDeps) :
+    addToDeps;
+}
+
+const allDependencies = flattenDeps([], ContainerFactory);
+
+// provide all dependencites to appInjector
+export const appInjector = allDependencies
+  .reduce((inj, factory) => inj.provide(factory, factory), injector());
 
 // Helper to inject custom components and return kepler.gl container
 export function injectComponents(recipes) {
