+++ src/components/side-panel.js
@@ -22,13 +22,13 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-import Sidebar from './side-panel/side-bar';
+import SidebarFactory from './side-panel/side-bar';
 import PanelHeaderFactory from './side-panel/panel-header';
-import LayerManager from './side-panel/layer-manager';
-import FilterManager from './side-panel/filter-manager';
-import InteractionManager from './side-panel/interaction-manager';
-import MapManager from './side-panel/map-manager';
-import PanelToggle from './side-panel/panel-toggle';
+import LayerManagerFactory from './side-panel/layer-manager';
+import FilterManagerFactory from './side-panel/filter-manager';
+import InteractionManagerFactory from './side-panel/interaction-manager';
+import MapManagerFactory from './side-panel/map-manager';
+import PanelToggleFactory from './side-panel/panel-toggle';
 
 import {
   ADD_DATA_ID,
@@ -48,7 +48,7 @@ const SidePanelContent = styled.div`
   overflow-x: hidden;
 `;
 
-const PanelTitle = styled.div`
+export const PanelTitleFactory = () => styled.div`
   color: ${props => props.theme.titleTextColor};
   font-size: 20px;
   font-weight: 400;
@@ -56,13 +56,31 @@ const PanelTitle = styled.div`
   margin-bottom: 14px;
 `;
 
-SidePanelFactory.deps = [PanelHeaderFactory];
+SidePanelFactory.deps = [
+  SidebarFactory,
+  PanelHeaderFactory,
+  PanelToggleFactory,
+  PanelTitleFactory,
+  LayerManagerFactory,
+  FilterManagerFactory,
+  InteractionManagerFactory,
+  MapManagerFactory
+];
 
 /**
  *
  * Vertical sidebar containing input components for the rendering layers
  */
-export default function SidePanelFactory(PanelHeader) {
+export default function SidePanelFactory(
+  Sidebar,
+  PanelHeader,
+  PanelToggle,
+  PanelTitle,
+  LayerManager,
+  FilterManager,
+  InteractionManager,
+  MapManager
+) {
 
   return class SidePanel extends Component {
     static propTypes = {
