+++ src/components/side-panel/layer-panel/layer-panel.js
@@ -36,103 +36,109 @@ const PanelWrapper = styled.div`
   }
 `;
 
-@sortable
-export default class LayerPanel extends Component {
-  static propTypes = {
-    layer: PropTypes.object.isRequired,
-    datasets: PropTypes.object.isRequired,
-    idx: PropTypes.number.isRequired,
-    layerConfigChange: PropTypes.func.isRequired,
-    layerTypeChange: PropTypes.func.isRequired,
-    openModal: PropTypes.func.isRequired,
-    removeLayer: PropTypes.func.isRequired,
-    onCloseConfig: PropTypes.func,
+function LayerPanelFactory() {
+  @sortable
+  class LayerPanel extends Component {
+    static propTypes = {
+      layer: PropTypes.object.isRequired,
+      datasets: PropTypes.object.isRequired,
+      idx: PropTypes.number.isRequired,
+      layerConfigChange: PropTypes.func.isRequired,
+      layerTypeChange: PropTypes.func.isRequired,
+      openModal: PropTypes.func.isRequired,
+      removeLayer: PropTypes.func.isRequired,
+      onCloseConfig: PropTypes.func,
 
-    layerTypeOptions: PropTypes.arrayOf(PropTypes.any),
-    layerVisConfigChange: PropTypes.func,
-    layerVisualChannelConfigChange: PropTypes.func
-  };
+      layerTypeOptions: PropTypes.arrayOf(PropTypes.any),
+      layerVisConfigChange: PropTypes.func,
+      layerVisualChannelConfigChange: PropTypes.func
+    };
 
-  updateLayerConfig = newProp => {
-    this.props.layerConfigChange(this.props.layer, newProp);
-  };
+    updateLayerConfig = newProp => {
+      this.props.layerConfigChange(this.props.layer, newProp);
+    };
 
-  updateLayerType = newType => {
-    this.props.layerTypeChange(this.props.layer, newType);
-  };
+    updateLayerType = newType => {
+      this.props.layerTypeChange(this.props.layer, newType);
+    };
 
-  updateLayerVisConfig = newVisConfig => {
-    this.props.layerVisConfigChange(this.props.layer, newVisConfig);
-  };
+    updateLayerVisConfig = newVisConfig => {
+      this.props.layerVisConfigChange(this.props.layer, newVisConfig);
+    };
 
-  updateLayerVisualChannelConfig = (newConfig, channel, scaleKey) => {
-    this.props.layerVisualChannelConfigChange(
-      this.props.layer,
-      newConfig,
-      channel,
-      scaleKey
-    );
-  };
+    updateLayerVisualChannelConfig = (newConfig, channel, scaleKey) => {
+      this.props.layerVisualChannelConfigChange(
+        this.props.layer,
+        newConfig,
+        channel,
+        scaleKey
+      );
+    };
 
-  _updateLayerLabel = ({target: {value}}) => {
-    this.updateLayerConfig({label: value});
-  };
+    _updateLayerLabel = ({target: {value}}) => {
+      this.updateLayerConfig({label: value});
+    };
 
-  _toggleVisibility = e => {
-    e.stopPropagation();
-    const isVisible = !this.props.layer.config.isVisible;
-    this.updateLayerConfig({isVisible});
-  };
+    _toggleVisibility = e => {
+      e.stopPropagation();
+      const isVisible = !this.props.layer.config.isVisible;
+      this.updateLayerConfig({isVisible});
+    };
 
-  _toggleEnableConfig = e => {
-    e.stopPropagation();
-    const {layer: {config: {isConfigActive}}} = this.props;
-    this.updateLayerConfig({isConfigActive: !isConfigActive});
-  };
+    _toggleEnableConfig = e => {
+      e.stopPropagation();
+      const {layer: {config: {isConfigActive}}} = this.props;
+      this.updateLayerConfig({isConfigActive: !isConfigActive});
+    };
 
-  _removeLayer = e => {
-    e.stopPropagation();
-    this.props.removeLayer(this.props.idx);
-  };
-  render() {
-    const {layer, idx, datasets, layerTypeOptions} = this.props;
-    const {config} = layer;
-    const {isConfigActive} = config;
+    _removeLayer = e => {
+      e.stopPropagation();
+      this.props.removeLayer(this.props.idx);
+    };
+    render() {
+      const {layer, idx, datasets, layerTypeOptions} = this.props;
+      const {config} = layer;
+      const {isConfigActive} = config;
 
-    return (
-      <PanelWrapper
-        active={isConfigActive}
-        className={`layer-panel ${this.props.className}`}
-        style={this.props.style}
-        onMouseDown={this.props.onMouseDown}
-        onTouchStart={this.props.onTouchStart}
-      >
-        <LayerPanelHeader
-          isConfigActive={isConfigActive}
-          id={layer.id}
-          idx={idx}
-          isVisible={config.isVisible}
-          label={config.label}
-          labelRCGColorValues={datasets[config.dataId].color}
-          layerType={layer.name}
-          onToggleEnableConfig={this._toggleEnableConfig}
-          onToggleVisibility={this._toggleVisibility}
-          onUpdateLayerLabel={this._updateLayerLabel}
-          onRemoveLayer={this._removeLayer}
-        />
-        {isConfigActive && (
-          <LayerConfigurator
-            layer={layer}
-            datasets={datasets}
-            layerTypeOptions={layerTypeOptions}
-            openModal={this.props.openModal}
-            updateLayerConfig={this.updateLayerConfig}
-            updateLayerVisualChannelConfig={this.updateLayerVisualChannelConfig}
-            updateLayerType={this.updateLayerType}
-            updateLayerVisConfig={this.updateLayerVisConfig}
+      return (
+        <PanelWrapper
+          active={isConfigActive}
+          className={`layer-panel ${this.props.className}`}
+          style={this.props.style}
+          onMouseDown={this.props.onMouseDown}
+          onTouchStart={this.props.onTouchStart}
+        >
+          <LayerPanelHeader
+            isConfigActive={isConfigActive}
+            id={layer.id}
+            idx={idx}
+            isVisible={config.isVisible}
+            label={config.label}
+            labelRCGColorValues={datasets[config.dataId].color}
+            layerType={layer.name}
+            onToggleEnableConfig={this._toggleEnableConfig}
+            onToggleVisibility={this._toggleVisibility}
+            onUpdateLayerLabel={this._updateLayerLabel}
+            onRemoveLayer={this._removeLayer}
           />
-        )}
-      </PanelWrapper>
-    );
+          {isConfigActive && (
+            <LayerConfigurator
+              layer={layer}
+              datasets={datasets}
+              layerTypeOptions={layerTypeOptions}
+              openModal={this.props.openModal}
+              updateLayerConfig={this.updateLayerConfig}
+              updateLayerVisualChannelConfig={this.updateLayerVisualChannelConfig}
+              updateLayerType={this.updateLayerType}
+              updateLayerVisConfig={this.updateLayerVisConfig}
+            />
+          )}
+        </PanelWrapper>
+      );
+    }
   }
+
+  return LayerPanel;
 }
+
+export default LayerPanelFactory;
