+++ src/components/side-panel/side-bar.js
@@ -52,73 +52,81 @@ const SideBarInner = styled.div`
   height: 100%;
 `;
 
-const CollapseButton = styled.div`
-  align-items: center;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
-  justify-content: center;
-  background-color: ${props => props.theme.sideBarCloseBtnBgd};
-  border-radius: 1px;
-  color: ${props => props.theme.sideBarCloseBtnColor};
-  display: flex;
-  height: 20px;
-  position: absolute;
-  right: -8px;
-  top: ${props => props.theme.sidePanel.margin.top}px;
-  width: 20px;
+export const CollapseButtonFactory = () => (
+  styled.div`
+    align-items: center;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+    justify-content: center;
+    background-color: ${props => props.theme.sideBarCloseBtnBgd};
+    border-radius: 1px;
+    color: ${props => props.theme.sideBarCloseBtnColor};
+    display: flex;
+    height: 20px;
+    position: absolute;
+    right: -8px;
+    top: ${props => props.theme.sidePanel.margin.top}px;
+    width: 20px;
 
-  :hover {
-    cursor: pointer;
-    box-shadow: none;
-    background-color: ${props => props.theme.sideBarCloseBtnBgdHover};
-  }
-`;
+    :hover {
+      cursor: pointer;
+      box-shadow: none;
+      background-color: ${props => props.theme.sideBarCloseBtnBgdHover};
+    }
+  `
+);
 
-export default class SideBar extends Component {
-  static defaultProps = {
-    width: 300,
-    minifiedWidth: 0,
-    isOpen: true,
-    onOpenOrClose: function noop() {}
+SidebarFactory.deps = [CollapseButtonFactory];
+
+function SidebarFactory(CollapseButton) {
+  return class SideBar extends Component {
+    static defaultProps = {
+      width: 300,
+      minifiedWidth: 0,
+      isOpen: true,
+      onOpenOrClose: function noop() {}
+    };
+
+    static propTypes = {
+      width: PropTypes.number,
+      isOpen: PropTypes.bool,
+      minifiedWidth: PropTypes.number,
+      onOpenOrClose: PropTypes.func
+    };
+
+    _onOpenOrClose = () => {
+      this.props.onOpenOrClose({isOpen: !this.props.isOpen});
+    };
+
+    render() {
+      const {isOpen, minifiedWidth, width} = this.props;
+      const horizontalOffset = isOpen ? 0 : minifiedWidth - width;
+
+      return (
+        <StyledSidePanelContainer
+          width={isOpen ? width : 0}
+          className="side-panel--container"
+        >
+          <SideBarContainer className="side-bar" style={{width: `${width}px`}}
+                            left={horizontalOffset}>
+            {isOpen ? (
+              <SideBarInner className="side-bar__inner">
+                {this.props.children}
+              </SideBarInner>
+            ) : null}
+            <CollapseButton
+              className="side-bar__close"
+              onClick={this._onOpenOrClose}
+            >
+              <ArrowRight
+                height="12px"
+                style={{transform: `rotate(${isOpen ? 180 : 0}deg)`}}
+              />
+            </CollapseButton>
+          </SideBarContainer>
+        </StyledSidePanelContainer>
+      );
+    }
   };
+}
 
-  static propTypes = {
-    width: PropTypes.number,
-    isOpen: PropTypes.bool,
-    minifiedWidth: PropTypes.number,
-    onOpenOrClose: PropTypes.func
-  };
-
-  _onOpenOrClose = () => {
-    this.props.onOpenOrClose({isOpen: !this.props.isOpen});
-  };
-
-  render() {
-    const {isOpen, minifiedWidth, width} = this.props;
-    const horizontalOffset = isOpen ? 0 : minifiedWidth - width;
-
-    return (
-      <StyledSidePanelContainer
-        width={isOpen ? width : 0}
-        className="side-panel--container"
-      >
-        <SideBarContainer className="side-bar" style={{width: `${width}px`}}
-                          left={horizontalOffset}>
-          {isOpen ? (
-            <SideBarInner className="side-bar__inner">
-              {this.props.children}
-            </SideBarInner>
-          ) : null}
-          <CollapseButton
-            className="side-bar__close"
-            onClick={this._onOpenOrClose}
-          >
-            <ArrowRight
-              height="12px"
-              style={{transform: `rotate(${isOpen ? 180 : 0}deg)`}}
-            />
-          </CollapseButton>
-        </SideBarContainer>
-      </StyledSidePanelContainer>
-    );
-  }
-};
+export default SidebarFactory;
