+++ src/components/side-panel/source-data-catalog.js
@@ -94,48 +94,52 @@ export const DatasetTag = ({onClick, dataset}) => (
   </DatasetTagWrapper>
 );
 
-const SourceDataCatalog = ({
-  datasets,
-  showDatasetTable,
-  removeDataset,
-  showDeleteDataset = false
-}) => (
-  <SourceDataCatelog className="source-data-catalog">
-    {Object.values(datasets).map((dataset, index) => (
-      <SidePanelSection key={dataset.id}>
-        <DatasetTitle className="source-data-title" clickable={Boolean(showDatasetTable)}>
-          <DatasetTag
-            dataset={dataset}
-            onClick={
-              showDatasetTable ? () => showDatasetTable(dataset.id) : null
-            }
-          />
-          {showDatasetTable ?
-            <CenterFlexbox className="source-data-arrow">
-              <ArrowRight height="12px" />
-            </CenterFlexbox> : null}
+function SourceDataCatalogFactory() {
+  const SourceDataCatalog = ({
+    datasets,
+    showDatasetTable,
+    removeDataset,
+    showDeleteDataset = false
+  }) => (
+    <SourceDataCatelog className="source-data-catalog">
+      {Object.values(datasets).map((dataset, index) => (
+        <SidePanelSection key={dataset.id}>
+          <DatasetTitle className="source-data-title" clickable={Boolean(showDatasetTable)}>
+            <DatasetTag
+              dataset={dataset}
+              onClick={
+                showDatasetTable ? () => showDatasetTable(dataset.id) : null
+              }
+            />
+            {showDatasetTable ?
+              <CenterFlexbox className="source-data-arrow">
+                <ArrowRight height="12px" />
+              </CenterFlexbox> : null}
+            {showDatasetTable ? (
+              <ShowDataTable
+                id={dataset.id}
+                showDatasetTable={showDatasetTable}
+              />
+            ) : null}
+            {showDeleteDataset ? (
+              <RemoveDataset
+                datasetKey={dataset.id}
+                removeDataset={removeDataset}
+              />
+            ) : null}
+          </DatasetTitle>
           {showDatasetTable ? (
-            <ShowDataTable
-              id={dataset.id}
-              showDatasetTable={showDatasetTable}
-            />
+            <DataRowCount className="source-data-rows">{`${numFormat(
+              dataset.allData.length
+            )} rows`}</DataRowCount>
           ) : null}
-          {showDeleteDataset ? (
-            <RemoveDataset
-              datasetKey={dataset.id}
-              removeDataset={removeDataset}
-            />
-          ) : null}
-        </DatasetTitle>
-        {showDatasetTable ? (
-          <DataRowCount className="source-data-rows">{`${numFormat(
-            dataset.allData.length
-          )} rows`}</DataRowCount>
-        ) : null}
-      </SidePanelSection>
-    ))}
-  </SourceDataCatelog>
-);
+        </SidePanelSection>
+      ))}
+    </SourceDataCatelog>
+  );
+
+  return SourceDataCatalog;
+}
 
 const ShowDataTable = ({id, showDatasetTable}) => (
   <DataTagAction
@@ -169,4 +173,4 @@ const RemoveDataset = ({datasetKey, removeDataset = defaultRemoveDataset}) => (
   </DataTagAction>
 );
 
-export default SourceDataCatalog;
+export default SourceDataCatalogFactory;
